requirements.txt
africastalking==1.2.9
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiohttp-retry==2.9.1
aiosignal==1.4.0
alembic==1.14.1
asgiref==3.8.1
attrs==25.3.0
bcrypt==4.3.0
bidict==0.23.1
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cryptography @ file:///home/builder/.termux-build/python-cryptography/src
Deprecated==1.2.18
distlib==0.3.9
dj-database-url==3.0.1
Django==5.2.3
django-cors-headers==4.7.0
django-extensions==4.1
django-redis==6.0.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
filelock==3.18.0
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Cache==0.13.1
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.10.1
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
frozenlist==1.7.0
greenlet==3.1.1
grep==0.3.2
gunicorn==23.0.0
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
limits==4.0.1
Mako==1.3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.6.3
numpy @ file:///home/builder/.termux-build/python-numpy/src/dist/numpy-2.2.5-cp312-cp312-linux_aarch64.whl#sha256=bb263430c58123c72e01dae41691aa54d2d399cc779d2e77fbf8b41564156a4b
ordered-set==4.1.0
packaging==24.2
panda==0.3.1
pandas @ file:///home/builder/.termux-build/python-pandas/src/dist/pandas-2.2.3-cp312-cp312-linux_aarch64.whl#sha256=e432b77b9c8f1bdb0448425ce1e03ac741b477fe9bcde42dab82cec19a797cee
pillow==11.1.0
pkg==0.2
platformdirs==4.3.6
propcache==0.3.2
psycopg2-binary==2.9.10
pycparser==2.22
Pygments==2.19.1
PyJWT==2.9.0
PyMySQL==1.1.1
pyotp==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.11.2
python-socketio==5.12.1
pytz==2025.2
redis==5.2.1
reportlab==4.3.0
requests==2.32.3
rest-framework-simplejwt==0.0.2
rich==13.9.4
schema==0.7.7
setuptools==80.9.0
simple-websocket==1.1.0
six==1.17.0
SQLAlchemy==2.0.37
sqlparse==0.5.3
tenacity==9.1.2
twilio==9.6.4
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
virtualenv==20.29.3
Werkzeug==3.1.3
wheel==0.45.1
whitenoise==6.9.0
wrapt==1.17.2
wsproto==1.2.0
xlsxwriter==3.2.5
yarl==1.20.1
